# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP SaaS API for Mauritanian businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies (super admin only)
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # USERS
...
  /roles/{id}/permissions:
    get:
      operationId: getRolePermissions
      tags: [roles]
      summary: Get permissions for a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role with permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleWithPermissions"
    put:
      operationId: setRolePermissions
      tags: [roles]
      summary: Set permissions for a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRolePermissionsBody"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleWithPermissions"

  # PARTNERS
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List partners
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [customer, supplier, both]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated partners
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPartners"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerBody"
      responses:
        "201":
          description: Created partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartnerBody"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PRODUCTS
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products and services
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product or service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # QUOTES
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        legalName:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        legalName:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
      required:
        - name
    UpdateCompanyBody:
      type: object
      properties:
        name:
          type: string
        legalName:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
...
    RoleWithPermissions:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - permissions
    SetRolePermissionsBody:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: string
      required:
        - permissions
    Partner:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        type:
          type: string
        name:
          type: string
        companyName:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - companyId
        - type
        - name
        - status
        - createdAt
        - updatedAt
    PaginatedPartners:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Partner"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreatePartnerBody:
      type: object
      properties:
        type:
          type: string
          enum: [customer, supplier, both]
        name:
          type: string
        companyName:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - type
        - name
    UpdatePartnerBody:
      type: object
      properties:
        type:
          type: string
          enum: [customer, supplier, both]
        name:
          type: string
        companyName:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        taxRate:
          type: number
        sku:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - companyId
        - name
        - type
        - unitPrice
        - taxRate
        - status
        - createdAt
        - updatedAt
    PaginatedProducts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [product, service]
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        taxRate:
          type: number
        sku:
          type: ["string", "null"]
      required:
        - name
        - type
        - unitPrice
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        taxRate:
          type: number
        sku:
          type: ["string", "null"]
        status:
          type: string
...