# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP SaaS API for Mauritanian businesses
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCompanyBody {
  name: string;
  /** @nullable */
  legalName?: string | null;
  /** @nullable */
  taxNumber?: string | null;
  /** @nullable */
  registrationNumber?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
}