# lib/api-zod/src/generated/types/paginatedPartners.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP SaaS API for Mauritanian businesses
 * OpenAPI spec version: 0.1.0
 */
import type { Partner } from "./partner";

export interface PaginatedPartners {
  data: Partner[];
  total: number;
  page: number;
  limit: number;
}